% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalR.R
\name{ear}
\alias{ear}
\title{Effective Annual Rate}
\usage{
ear(apr, n, p = 5)
}
\arguments{
\item{apr}{Annual Rate (Nominal Interest Rate)}

\item{n}{Number of compounds in a year}

\item{p}{Calculates the EAR to the (1/10^p) decimal place}
}
\value{
Effective Annual Rate
}
\description{
Effective Annual Rate
}
\examples{
ear(apr= 0.05, n = 12)
}
