% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalR.R
\name{fv}
\alias{fv}
\title{Future Value}
\usage{
fv(pv, r, n)
}
\arguments{
\item{pv}{Present Value}

\item{r}{Discount Rate}

\item{n}{Number of Compounding Periods}
}
\value{
Returns the Future Value
}
\description{
Calculates the Future Value given a Present Value
}
\examples{
fv(5000, 0.08/12, 5*12)
}
