% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalR.R
\name{geometric}
\alias{geometric}
\title{Geometric Mean Return}
\usage{
geometric(c)
}
\arguments{
\item{c}{Periodic returns in decimal form}
}
\value{
Returns the Geometric Mean Return
}
\description{
Geometric Mean Return
}
\examples{
geometric(c(0.05, 0.02, -0.03, 0.09, -0.02))
}
