% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalR.R
\name{ipmt}
\alias{ipmt}
\title{Interest Payment}
\usage{
ipmt(pv, fv = 0, n, r, x, end = TRUE)
}
\arguments{
\item{pv}{Present Value}

\item{fv}{Future Value}

\item{n}{Number of Periods}

\item{r}{Rate}

\item{x}{Period in which to calculate the interest portion of the payment}

\item{end}{If FALSE, payments are made at the beginning of the period}
}
\value{
Returns the Interest Portion of the Payment in Period "x"
}
\description{
Calculates the interest portion of the payment in period "x"
}
\examples{
ipmt(pv = 20000, fv = 0, n = 5 * 12, r = 0.05/12, x = 12, end = TRUE)
}
