% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalR.R
\name{pos}
\alias{pos}
\title{Position}
\usage{
pos(df, n)
}
\arguments{
\item{df}{Dataframe}

\item{n}{Number}
}
\value{
Returns the General Position
}
\description{
Position
}
\examples{
\dontrun{ pos(dataframe, n)}
}
