% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalR.R
\name{pv}
\alias{pv}
\title{Present Value}
\usage{
pv(fv, r, n)
}
\arguments{
\item{fv}{Future Value}

\item{r}{Discount Rate}

\item{n}{Number of Compounding Periods}
}
\value{
Returns the Present Value
}
\description{
Calculates the present value of a given future value
}
\examples{
pv(5000, 0.08/12, 5*12)
}
