% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalR.R
\name{r.calc}
\alias{r.calc}
\title{Return Calculation}
\usage{
r.calc(vector)
}
\arguments{
\item{vector}{Vector from which to calculate the periodic return}
}
\value{
Returns the Periodic Percent Return
}
\description{
Return Calculation
}
\examples{
r.calc(c(100, 75, 50, 80, 125))
}
