% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalR.R
\name{schedule}
\alias{schedule}
\title{Amortization Schedule}
\usage{
schedule(r, n, pv, fv = 0, end = TRUE)
}
\arguments{
\item{r}{Rate}

\item{n}{Number of Periods}

\item{pv}{Present Value}

\item{fv}{Future Value, set = 0}

\item{end}{If FALSE, payments are made at the beginning of the period}
}
\value{
Returns the Amortization Schedule in a dataframe
}
\description{
Creates an amortization schedule of a loan
}
\examples{
schedule(r = 0.06/12, n = 10 * 12, pv = -5000, fv = 0, end = TRUE)
}
